PSTH quick fixes
TTL inputs generate Events at their rising and at their falling edges. In the EvntTrigAvg plugin, TTL are then counted twice and cause the PSTH to be wrong. This is particularly obvious for long TTLs (such as light stimulation lasting for up to few hundreds of ms). Additionnaly, a bug for  recording with more than 16 electrodes was corrected.